refactor(currency): move FormatNumber and FormatNumberByParts into currency.go as struct methods
